2to3-assisted Py3 migrations
Used `2to3 -x future -x unicode halonctl` to scan for Py3-incompatible constructs, and manually converted them to use `six` instead.

+++ halonctl/modules/queue.py
@@ -21,7 +21,7 @@ class QueueModule(Module):
 				self.partial = True
 				continue
 			
-			count = long(result[1].all().strip().split('=')[1])
+			count = int(result[1].all().strip().split('=')[1])
 			totalCount += count
 			
 			if not args.quiet:
